@@ -5,8 +5,8 @@ author: jlcool <jiangliang.net@gmail.com>
 homepage: https://github.com/jlcool/aliossflutter
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
-  flutter: ">=1.5.9-pre.94 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=1.5.9-pre.94"
 
 dependencies:
   uuid: ^3.0.4
