@@ -1,49 +1,49 @@
 class ProgressResponse {
-  final double currentSize;
-  final String key;
-  final double totalSize;
+  final double? currentSize;
+  final String? key;
+  final double? totalSize;
 
   ProgressResponse({this.currentSize, this.totalSize,this.key});
 
   double getProgress() {
     // TODO: implement toString
-    return currentSize / totalSize;
+    return currentSize! / totalSize!;
   }
 }
 class SignResponse {
-  bool success;
-  String url;
-  String key;
-  String msg;
+  bool? success;
+  String? url;
+  String? key;
+  String? msg;
   SignResponse({this.success, this.url,this.msg,this.key});
 }
 class UploadResponse {
-    bool success;
-    String key;
-    String msg;
-    String servercallback;
+    bool? success;
+    String? key;
+    String? msg;
+    String? servercallback;
   UploadResponse({this.success, this.key,this.msg,this.servercallback});
 }
 class DownloadResponse {
-    bool success;
-    String path;
-    String key;
-    String msg;
+    bool? success;
+    String? path;
+    String? key;
+    String? msg;
   DownloadResponse({this.success, this.path,this.msg,this.key});
 }
 class DeleteResponse {
-    bool success;
-    String key;
+    bool? success;
+    String? key;
   DeleteResponse({this.success,this.key});
 }
 class HeadObjectResponse {
-  bool success;
-  int lastModified;
-  String key;
+  bool? success;
+  int? lastModified;
+  String? key;
   HeadObjectResponse({this.success,this.lastModified,this.key});
 }
 class ListObjectsResponse {
-    bool success;
-    List objects;
+    bool? success;
+    List? objects;
   ListObjectsResponse({this.success,this.objects});
 }
